Move lsf test to lsf test file
